Deploy the configs for the Directory
This includes the permissions for the new alerts API.

Many pieces cannot be deployed here yet as they rely on the SA UUID. I
need to decide whether to stick with Service Requirements or whether to
try to create dynamic dump files. In any case for the Directory SA in
particular we have to be careful about bootstrapping.